update: command groups + permissions

+++ commands/notes.js
@@ -1,4 +1,3 @@
-const Discord = require("discord.js");
 const { creatorID } = require("../config.json");
 const https = require("https");
 let fs = require("fs");
@@ -8,13 +7,10 @@ module.exports = {
   description: "notes for subjects",
   aliases: ["poznamky"],
   args: 1,
-  usage: [
-    "[subject]  => get file for that subject ",
-    "files  => list all subjects with files",
-    "subjects  => list all available subjects",
-    "[subject] + file  => save file under this subject",
-  ],
+  usage: ["[subject]", "list", "subjects"],
   requirePermission: true,
+  group: "school",
+  DM: true,
   execute(message, args) {
     const subjectName = args[0].toUpperCase();
     let data = fs.readFileSync("./notes/notes.json");
@@ -31,11 +27,7 @@ module.exports = {
           subjectArr.push(`${key}: ${subjects[key]}`);
         }
       }
-      return message.reply(
-        `here's a list of all subjects and their files:\n${subjectArr.join(
-          "\n"
-        )}`
-      );
+      return message.channel.send(`here's a list of all subjects and their files:\n\`${subjectArr.join("\n")}\``);
     }
     if (args[0].toLowerCase() == "subjects") {
       // list all subjects with their files
@@ -45,14 +37,9 @@ module.exports = {
           subjectArr.push(`${key}: ${subjects[key]}`);
         }
       }
-      return message.reply(
-        `here's a list of all subjects: \n ${Object.keys(subjects).join(", ")}`
-      );
+      return message.channel.send(`here's a list of all subjects: \n\`${Object.keys(subjects).join(", ")}\``);
     }
-    if (
-      args[0].toLowerCase() == "remove" &&
-      message.author.id == creatorID[0]
-    ) {
+    if (args[0].toLowerCase() == "remove" && message.author.id == creatorID[0]) {
       let filter = (m) => m.author.id === message.author.id;
       message.channel.send("subject name: ").then(() => {
         message.channel
@@ -69,9 +56,7 @@ module.exports = {
               fs.unlinkSync(`notes/${subjects[message.content.toUpperCase()]}`);
 
               subjects[message.content.toUpperCase()] = ""; // delete file path
-              // save data back to file
               data = JSON.stringify(subjects);
-              // write JSON string to a file
               fs.writeFile("./notes/notes.json", data, (err) => {
                 if (err) {
                   throw err;
@@ -79,7 +64,7 @@ module.exports = {
               });
               return message.channel.send("file successfully removed");
             } else {
-              return message.reply("there isn't any file for this subject");
+              return message.channel.send("there isn't any file for this subject");
             }
           })
           .catch((collected) => {
@@ -87,58 +72,36 @@ module.exports = {
           });
       });
     } else {
-      //! only creator can upload files (yes, only me)
+      //! only creator can upload files
       if (message.author.id == creatorID[0]) {
         if (message.attachments.first()) {
           const nameOfFile = message.attachments.first().filename;
-          if (
-            supportedFileTypes.includes(
-              nameOfFile.slice(nameOfFile.indexOf("."))
-            )
-          ) {
-            // message.channel.send("file type supported");
-          } else {
-            return message.channel.send(
-              `I'm sorry this file type is not supported`
-            );
-          }
+          if (!supportedFileTypes.includes(nameOfFile.slice(nameOfFile.indexOf("."))))
+            return message.channel.send(`I'm sorry, this file type is not supported`);
+
           if (subjects[subjectName] != "") {
             fs.unlinkSync(`./notes/${subjects[subjectName]}`); // remove old file before downloading new one (if there is any)
           }
           const file = fs.createWriteStream(`./notes/${nameOfFile}`);
-          const request = https.get(
-            message.attachments.first().url,
-            function (response) {
-              response.pipe(file);
-            }
-          );
+          const request = https.get(message.attachments.first().url, function (response) {
+            response.pipe(file);
+          });
           if (subjects[subjectName]) {
             // check if subjects, we want to save data as exists
             subjects[subjectName] = nameOfFile; // save file name into subject (this way we can find them later and keep their original name)
           } else {
-            subjects[subjectName] = [].concat(nameOfFile); // this should not happen (subjects are predefined)
+            subjects[subjectName] = [].concat(nameOfFile);
           }
-          // save data back to file
           data = JSON.stringify(subjects);
-          // write JSON string to a file
           fs.writeFile("./notes/notes.json", data, (err) => {
             if (err) {
               throw err;
             }
           });
-          // then return
-          return message.channel.send(
-            `file: ${nameOfFile} successfully uploaded to ${subjectName} subject`
-          );
+          return message.channel.send(`file: \`${nameOfFile}\` successfully uploaded to \`${subjectName}\` subject`);
         }
       }
       if (subjects[subjectName]) {
-        //! important, files will be sent to dm from now on, it's way more secure and files will get only to the users with permissions who will value our service
-        // since now all subjects are declared we just need to check if subject has some file property
-        // send file to channel (old)
-        // return message.reply(subjectName + ":", {
-        //   files: [`./notes/${subjects[subjectName]}`],
-        // });
         return message.author
           .send(
             `here is your file for: ${subjectName}\n\`please don't share !\` (you don't want to get hurt by my 15 in guns)\nenjoy :smile:`,
@@ -149,19 +112,13 @@ module.exports = {
           )
           .then(() => {
             if (message.channel.type === "dm") return;
-            message.reply("I've sent you a DM");
+            message.channel.send("I've sent you a DM");
           })
-          .catch((error) => {
-            console.error(
-              `Could not send help DM to ${message.author.tag}.\n`,
-              error
-            );
-            message.reply(
-              "it seems like I can't DM you! Do you have DMs disabled?"
-            );
+          .catch(() => {
+            message.reply("it seems like I can't DM you! Do you have DMs disabled?");
           });
       } else {
-        return message.reply("Sorry, I don't have any file for this subject");
+        return message.channel.send("Sorry, I don't have any file for this subject");
       }
     }
   },
